requirements.txt
uvicorn>=0.15.0
fastapi>=0.68.0
SQLALchemy>=1.4.0
alembic>=1.10.0
python-dotenv>=0.19.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-jose[cryptography]>=3.3.0
openai==1.76
email-validator==2.1.0.post1
python-multipart>=0.0.6
openai-whisper
passlib[bcrypt]>=1.7.4
bcrypt==4.3.0
boto3==1.34.59
aiofiles==23.2.1
pydantic>=1.8.0
SpeechRecognition>=3.8.1
pydub>=0.25.1
faster-whisper>=0.9.0
torch>=2.0.0
numpy>=1.21.0
ffmpeg-python>=0.2.0
jieba>=0.42.1
regex>=2023.0.0
azure-cognitiveservices-speech>=1.31.0
fastapi-mail==1.4.1
jinja2==3.1.2